+++ packages/core-js/full/set/difference.js
@@ -1,5 +1,5 @@
+require('../../actual/set/difference');
 require('../../modules/es.array.iterator');
-require('../../modules/es.set');
 require('../../modules/es.string.iterator');
 require('../../modules/esnext.set.difference');
 require('../../modules/web.dom-collections.iterator');