+++ packages/core-js/full/set/is-subset-of.js
@@ -1,5 +1,5 @@
+require('../../actual/set/is-subset-of');
 require('../../modules/es.array.iterator');
-require('../../modules/es.set');
 require('../../modules/es.string.iterator');
 require('../../modules/esnext.set.is-subset-of');
 require('../../modules/web.dom-collections.iterator');