+++ packages/core-js/full/set/union.js
@@ -1,5 +1,5 @@
+require('../../actual/set/union');
 require('../../modules/es.array.iterator');
-require('../../modules/es.set');
 require('../../modules/es.string.iterator');
 require('../../modules/esnext.set.union');
 require('../../modules/web.dom-collections.iterator');