+++ packages/core-js/internals/set-union.js
@@ -1,0 +1,20 @@
+'use strict';
+var SetHelpers = require('../internals/set-helpers');
+var getSetRecord = require('../internals/get-set-record');
+var iterateSimple = require('../internals/iterate-simple');
+
+var aSet = SetHelpers.aSet;
+var add = SetHelpers.add;
+var clone = SetHelpers.clone;
+
+// `Set.prototype.union` method
+// https://github.com/tc39/proposal-set-methods
+module.exports = function union(other) {
+  var O = aSet(this);
+  var keysIter = getSetRecord(other).getIterator();
+  var result = clone(O);
+  iterateSimple(keysIter, function (it) {
+    add(result, it);
+  });
+  return result;
+};