+++ packages/core-js/modules/esnext.set.difference.js
@@ -1,22 +1,14 @@
 'use strict';
 var $ = require('../internals/export');
-var getBuiltIn = require('../internals/get-built-in');
 var call = require('../internals/function-call');
-var aCallable = require('../internals/a-callable');
-var anObject = require('../internals/an-object');
-var speciesConstructor = require('../internals/species-constructor');
-var iterate = require('../internals/iterate');
+var toSetLike = require('../internals/to-set-like');
+var $difference = require('../internals/set-difference');
 
 // `Set.prototype.difference` method
 // https://github.com/tc39/proposal-set-methods
+// TODO: Obsolete version, remove from `core-js@4`
 $({ target: 'Set', proto: true, real: true, forced: true }, {
-  difference: function difference(iterable) {
-    var set = anObject(this);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))(set);
-    var remover = aCallable(newSet['delete']);
-    iterate(iterable, function (value) {
-      call(remover, newSet, value);
-    });
-    return newSet;
+  difference: function difference(other) {
+    return call($difference, this, toSetLike(other));
   }
 });