+++ packages/core-js/modules/esnext.set.intersection.js
@@ -1,23 +1,14 @@
 'use strict';
 var $ = require('../internals/export');
-var getBuiltIn = require('../internals/get-built-in');
 var call = require('../internals/function-call');
-var aCallable = require('../internals/a-callable');
-var anObject = require('../internals/an-object');
-var speciesConstructor = require('../internals/species-constructor');
-var iterate = require('../internals/iterate');
+var toSetLike = require('../internals/to-set-like');
+var $intersection = require('../internals/set-intersection');
 
 // `Set.prototype.intersection` method
 // https://github.com/tc39/proposal-set-methods
+// TODO: Obsolete version, remove from `core-js@4`
 $({ target: 'Set', proto: true, real: true, forced: true }, {
-  intersection: function intersection(iterable) {
-    var set = anObject(this);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))();
-    var hasCheck = aCallable(set.has);
-    var adder = aCallable(newSet.add);
-    iterate(iterable, function (value) {
-      if (call(hasCheck, set, value)) call(adder, newSet, value);
-    });
-    return newSet;
+  intersection: function intersection(other) {
+    return call($intersection, this, toSetLike(other));
   }
 });