+++ packages/core-js/modules/esnext.set.is-subset-of.js
@@ -1,26 +1,14 @@
 'use strict';
 var $ = require('../internals/export');
-var getBuiltIn = require('../internals/get-built-in');
 var call = require('../internals/function-call');
-var aCallable = require('../internals/a-callable');
-var isCallable = require('../internals/is-callable');
-var anObject = require('../internals/an-object');
-var getIterator = require('../internals/get-iterator');
-var iterate = require('../internals/iterate');
+var toSetLike = require('../internals/to-set-like');
+var $isSubsetOf = require('../internals/set-is-subset-of');
 
 // `Set.prototype.isSubsetOf` method
-// https://tc39.github.io/proposal-set-methods/#Set.prototype.isSubsetOf
+// https://github.com/tc39/proposal-set-methods
+// TODO: Obsolete version, remove from `core-js@4`
 $({ target: 'Set', proto: true, real: true, forced: true }, {
-  isSubsetOf: function isSubsetOf(iterable) {
-    var iterator = getIterator(this);
-    var otherSet = anObject(iterable);
-    var hasCheck = otherSet.has;
-    if (!isCallable(hasCheck)) {
-      otherSet = new (getBuiltIn('Set'))(iterable);
-      hasCheck = aCallable(otherSet.has);
-    }
-    return !iterate(iterator, function (value, stop) {
-      if (call(hasCheck, otherSet, value) === false) return stop();
-    }, { IS_ITERATOR: true, INTERRUPTED: true }).stopped;
+  isSubsetOf: function isSubsetOf(other) {
+    return call($isSubsetOf, this, toSetLike(other));
   }
 });