+++ packages/core-js/modules/esnext.set.symmetric-difference.js
@@ -1,23 +1,14 @@
 'use strict';
 var $ = require('../internals/export');
-var getBuiltIn = require('../internals/get-built-in');
 var call = require('../internals/function-call');
-var aCallable = require('../internals/a-callable');
-var anObject = require('../internals/an-object');
-var speciesConstructor = require('../internals/species-constructor');
-var iterate = require('../internals/iterate');
+var toSetLike = require('../internals/to-set-like');
+var $symmetricDifference = require('../internals/set-symmetric-difference');
 
 // `Set.prototype.symmetricDifference` method
 // https://github.com/tc39/proposal-set-methods
+// TODO: Obsolete version, remove from `core-js@4`
 $({ target: 'Set', proto: true, real: true, forced: true }, {
-  symmetricDifference: function symmetricDifference(iterable) {
-    var set = anObject(this);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))(set);
-    var remover = aCallable(newSet['delete']);
-    var adder = aCallable(newSet.add);
-    iterate(iterable, function (value) {
-      call(remover, newSet, value) || call(adder, newSet, value);
-    });
-    return newSet;
+  symmetricDifference: function symmetricDifference(other) {
+    return call($symmetricDifference, this, toSetLike(other));
   }
 });