+++ tests/unit-pure/esnext.set.intersection.js
@@ -1,4 +1,4 @@
-import { createIterable } from '../helpers/helpers';
+import { createIterable, createSetLike } from '../helpers/helpers';
 
 import from from 'core-js-pure/es/array/from';
 import Set from 'core-js-pure/full/set';
@@ -12,8 +12,21 @@ QUnit.test('Set#intersection', assert => {
   assert.nonEnumerable(Set.prototype, 'intersection');
 
   const set = new Set([1]);
-  assert.notSame(set.intersection([2]), set);
+  assert.notSame(set.intersection(new Set()), set);
 
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(new Set([4, 5]))), []);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(new Set([2, 3, 4]))), [2, 3]);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(createSetLike([4, 5]))), []);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(createSetLike([2, 3, 4]))), [2, 3]);
+
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(new Set([3, 2]))), [2, 3]);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(new Set([3, 2, 1]))), [1, 2, 3]);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(new Set([3, 2, 1, 0]))), [1, 2, 3]);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(createSetLike([3, 2]))), [2, 3]);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(createSetLike([3, 2, 1]))), [1, 2, 3]);
+  assert.deepEqual(from(new Set([1, 2, 3]).intersection(createSetLike([3, 2, 1, 0]))), [1, 2, 3]);
+
+  // TODO: drop from core-js@4
   assert.deepEqual(from(new Set([1, 2, 3]).intersection([4, 5])), []);
   assert.deepEqual(from(new Set([1, 2, 3]).intersection([2, 3, 4])), [2, 3]);
   assert.deepEqual(from(new Set([1, 2, 3]).intersection(createIterable([2, 3, 4]))), [2, 3]);