+++ tests/unit-pure/esnext.set.is-disjoint-from.js
@@ -1,4 +1,4 @@
-import { createIterable } from '../helpers/helpers';
+import { createIterable, createSetLike } from '../helpers/helpers';
 
 import Set from 'core-js-pure/full/set';
 
@@ -10,11 +10,20 @@ QUnit.test('Set#isDisjointFrom', assert => {
   assert.name(isDisjointFrom, 'isDisjointFrom');
   assert.nonEnumerable(Set.prototype, 'isDisjointFrom');
 
+  assert.true(new Set([1]).isDisjointFrom(new Set([2])));
+  assert.false(new Set([1]).isDisjointFrom(new Set([1])));
+  assert.true(new Set([1, 2, 3]).isDisjointFrom(new Set([4, 5, 6])));
+  assert.false(new Set([1, 2, 3]).isDisjointFrom(new Set([5, 4, 3])));
+  assert.true(new Set([1]).isDisjointFrom(createSetLike([2])));
+  assert.false(new Set([1]).isDisjointFrom(createSetLike([1])));
+  assert.true(new Set([1, 2, 3]).isDisjointFrom(createSetLike([4, 5, 6])));
+  assert.false(new Set([1, 2, 3]).isDisjointFrom(createSetLike([5, 4, 3])));
+
+  // TODO: drop from core-js@4
   assert.true(new Set([1]).isDisjointFrom([2]));
   assert.false(new Set([1]).isDisjointFrom([1]));
   assert.true(new Set([1, 2, 3]).isDisjointFrom([4, 5, 6]));
   assert.false(new Set([1, 2, 3]).isDisjointFrom([5, 4, 3]));
-
   assert.true(new Set([1]).isDisjointFrom(createIterable([2])));
   assert.false(new Set([1]).isDisjointFrom(createIterable([1])));